refactor(history-api-fallback): split from serve
