+++ packages/svrx/lib/constant.js
@@ -27,7 +27,7 @@ module.exports = {
   // REMOVE POSTFIX like `-beta` in  0.0.1-beta.
   /* eslint-disable global-require, import/no-dynamic-require */
   VERSION: require(libPath.join(__dirname, '../package.json')).version.replace(/-.*$/, ''),
-  BUILTIN_PLUGIN: ['livereload', 'proxy', 'serve', 'cors', 'open'],
+  BUILTIN_PLUGIN: ['livereload', 'proxy', 'serve', 'cors', 'open', 'history-api-fallback'],
   CUSTOM_SCHEMA_TYPES: ['compute', 'function'],
   GROUPS,
   EVENTS: {
